[container.vector-interface] move is_vector to type_traits.hh
