+++ dune/xt/la/solver/istl.hh
@@ -24,7 +24,6 @@
 
 #include <dune/xt/common/exceptions.hh>
 #include <dune/xt/common/configuration.hh>
-#include <dune/xt/common/memory.hh>
 
 #include <dune/xt/la/container/istl.hh>
